+++ src/main/java/com/even/zaro/controller/ProfileController.java
@@ -2,18 +2,18 @@ package com.even.zaro.controller;
 
 import com.even.zaro.dto.UserPostDto;
 import com.even.zaro.dto.UserProfileDto;
-import com.even.zaro.dto.favoriteDTO.FavoriteAddRequest;
-import com.even.zaro.dto.favoriteDTO.FavoriteAddResponse;
-import com.even.zaro.dto.favoriteDTO.FavoriteEditRequest;
-import com.even.zaro.dto.favoriteDTO.FavoriteResponse;
+import com.even.zaro.dto.favorite.FavoriteAddRequest;
+import com.even.zaro.dto.favorite.FavoriteAddResponse;
+import com.even.zaro.dto.favorite.FavoriteEditRequest;
+import com.even.zaro.dto.favorite.FavoriteResponse;
 import com.even.zaro.global.ApiResponse;
 import com.even.zaro.service.ProfileService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import com.even.zaro.dto.profileDTO.GroupCreateRequest;
-import com.even.zaro.dto.profileDTO.GroupEditRequest;
-import com.even.zaro.dto.profileDTO.GroupResponse;
+import com.even.zaro.dto.profile.GroupCreateRequest;
+import com.even.zaro.dto.profile.GroupEditRequest;
+import com.even.zaro.dto.profile.GroupResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
