+++ src/main/java/com/even/zaro/dto/favorite/FavoriteAddResponse.java
@@ -1,4 +1,4 @@
-package com.even.zaro.dto.favoriteDTO;
+package com.even.zaro.dto.favorite;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
