+++ src/main/java/com/even/zaro/dto/favorite/FavoriteEditRequest.java
@@ -1,7 +1,6 @@
-package com.even.zaro.dto.favoriteDTO;
+package com.even.zaro.dto.favorite;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
 import lombok.Data;
 
 @Data
