+++ src/main/java/com/even/zaro/dto/profile/GroupEditRequest.java
@@ -1,9 +1,7 @@
-package com.even.zaro.dto.profileDTO;
+package com.even.zaro.dto.profile;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-
-import java.time.LocalDateTime;
 
 @Data
 public class GroupEditRequest {
