+++ src/main/java/com/even/zaro/dto/profile/GroupResponse.java
@@ -1,4 +1,4 @@
-package com.even.zaro.dto.profileDTO;
+package com.even.zaro.dto.profile;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
