+++ src/main/java/com/even/zaro/service/ProfileService.java
@@ -2,13 +2,13 @@ package com.even.zaro.service;
 
 import com.even.zaro.dto.UserPostDto;
 import com.even.zaro.dto.UserProfileDto;
-import com.even.zaro.dto.favoriteDTO.FavoriteAddRequest;
-import com.even.zaro.dto.favoriteDTO.FavoriteAddResponse;
-import com.even.zaro.dto.favoriteDTO.FavoriteEditRequest;
-import com.even.zaro.dto.favoriteDTO.FavoriteResponse;
-import com.even.zaro.dto.profileDTO.GroupCreateRequest;
-import com.even.zaro.dto.profileDTO.GroupEditRequest;
-import com.even.zaro.dto.profileDTO.GroupResponse;
+import com.even.zaro.dto.favorite.FavoriteAddRequest;
+import com.even.zaro.dto.favorite.FavoriteAddResponse;
+import com.even.zaro.dto.favorite.FavoriteEditRequest;
+import com.even.zaro.dto.favorite.FavoriteResponse;
+import com.even.zaro.dto.profile.GroupCreateRequest;
+import com.even.zaro.dto.profile.GroupEditRequest;
+import com.even.zaro.dto.profile.GroupResponse;
 import com.even.zaro.entity.Favorite;
 import com.even.zaro.entity.FavoriteGroup;
 import com.even.zaro.entity.Place;
